requirements.txt
firebase-admin==6.0.0
python-dotenv==0.21.1
PyQt5==5.15.9
